backend/requirements.txt
Werkzeug==0.16.1

python-dotenv>=0.15.0,<0.16.0
Flask>=1.1.2,<1.2.0
Flask-RESTful>=0.3.8,<0.4.0
Flask-BabelPlus>=2.2.0,<2.3.0
Flask-SQLAlchemy>=2.4.4,<2.5.0
flask-marshmallow>=0.14.0,<0.15.0
marshmallow-sqlalchemy>=0.24.1,<0.25.0
